tasker/serializers/task_serializer.py
from rest_framework import serializers
from tasker.model.task import Task

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = '__all__'    

        